# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Celsius API",
    "version": "1.0",
    "description": "Celsius Mobile API Documentation"
  },
  "servers": [
    {
      "url": "https://celsiuspop.com",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/v1/wishlist/{productId}": {
      "delete": {
        "summary": "Remove a product from wishlist",
        "tags": [
          "Wishlist"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Removed from wishlist"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Wishlist not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/webhooks/shipquickr": {
      "post": {
        "summary": "Shipquickr Webhook",
        "description": "Receives order status updates and AWB tracking info from Shipquickr",
        "tags": [
          "Webhooks"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "orderId",
                  "event"
                ],
                "properties": {
                  "orderId": {
                    "type": "string"
                  },
                  "event": {
                    "type": "string"
                  },
                  "awb": {
                    "type": "string"
                  },
                  "courierUrl": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook processed successfully"
          },
          "400": {
            "description": "Missing required fields"
          },
          "404": {
            "description": "Order not found"
          },
          "500": {
            "description": "Internal Server Error processing webhook"
          }
        }
      }
    }
  },
...
}